% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{between}
\alias{between}
\title{Detect if values fall within a range}
\usage{
between(x, left, right, inclusive = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{left, right}{Numeric boundary values. Both \code{left} and \code{right} are
recycled to the size of \code{x}.}

\item{inclusive}{A logical to specify that \code{left} and \code{right} boundary
values should be accepted in the range (the \emph{default}) or excluded if
\code{FALSE}.}
}
\value{
A logical vector the same length as \code{x}.
}
\description{
Vectorised inclusive or exclusive \code{x >= left} & \code{x <= right}
}
\details{
\code{inclusive = FALSE} could be used to test for positive non-zero values:
\code{between(x, 0, Inf, inclusive = FALSE)}.
}
\seealso{
\code{\link[dplyr:between]{dplyr::between()}}
}
\keyword{internal}
