% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataConvert.R
\name{dataConvert}
\alias{dataConvert}
\title{Converting data into format used for the ExplodeLayout(EL) Algorithm}
\usage{
dataConvert(net_matrix, cluster, outcome, entity)
}
\arguments{
\item{net_matrix}{a numeric matrix or dataframe storing the network as adjacency matrix(for unipartite networks)
or incidence matrix(for bipartite networks). It would be great if you have vertices names contained.}

\item{cluster}{a numeric vector containing the cluster membership for each vertices. Make sure the length
matches the data matrix file.}

\item{outcome}{a numeric vector containing the outcome for each vertice(such as 1 for cases and 2 for controls)}

\item{entity}{a numeric vector representing the entity for each vertice (if it's unipartite, put all 1, if its'
bipartite network, put one type vertice as 1, the other as 2)}
}
\description{
This function takes user's data, including the adjacency or incidence matrix representing networks,
and vectors containing information of the vertices' cluster membership, entity, etc.
}
\details{
The output of the data would be a list containing two dataframes, 'nodelist' and 'net', which would
serve as the input of the next 'search' step for EL algorithm.
}
