% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epl-package.R
\docType{package}
\name{epl-package}
\alias{epl}
\alias{epl-package}
\title{epl: Functions and Scripts for the UBC Environmental Physiology Laboratory}
\description{
A repository of functions and scripts for common tasks for use by students in the UBC Environmental Physiology Laboratory.
}
\author{
\strong{Maintainer}: Jem Arnold \email{jem.arnold@ubc.ca}

}
\keyword{internal}
