% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epl.R
\docType{package}
\name{epl}
\alias{epl}
\alias{epl-package}
\title{epl:A package to implement the ExplodeLayout algorithm for networks.}
\description{
The epl package provides one function to apply the ExplodeLayout algorithm
for network visuallization
}
\section{explodelayout functions}{

The function takes as input two parameters....
}

