% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_helpers.R
\name{example_epl}
\alias{example_epl}
\title{Get path to \code{{epl}} example files}
\usage{
example_epl(file = NULL)
}
\arguments{
\item{file}{Name of file as character string. If kept as \code{NULL}, returns a
vector of all available files.}
}
\value{
File paths for selected example files stored in this package.
}
\description{
Get path to \code{{epl}} example files
}
\examples{
example_epl()
example_epl("tymewear_live")

}
