% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Find peak values from rows within a given timespan}
\usage{
find_peaks(data, x, y, span = 30, between = NULL)
}
\arguments{
\item{data}{A data frame with at least numeric \code{x} and \code{y} columns.}

\item{x}{A character string for the column name of \code{x} (e.g. \emph{"time"}.
Must match column name exactly.}

\item{y}{A character string for the column name of \code{y} (e.g. \emph{"VO2"}.
Must match column name exactly.}

\item{span}{A numeric value defining timespan in which to calculate rolling
local means, from which peak values will be taken. \code{span = 0} will take
the single highest sample.}

\item{between}{An optional time range in the form \code{c(min, max)} in units of
\code{x}, within which to look for peak values (inclusive).}
}
\value{
A \link[tibble:tibble-package]{tibble} with one row of mean values
from all columns across the peak-detected rows. \code{Samples} column shows
the number of samples within the timespan.
}
\description{
Detects the peak values of \code{y} column within a \code{span} of \code{x} column (e.g.
timespan), returns the mean of all columns within those rows.
}
\details{
The function uses a rolling window approach:
\enumerate{
\item For each non-NA value of \code{y}, calculates the mean of \code{y} within a
window from time \code{t} to \code{t + span}.
\item Identifies the window with the maximum mean value of \code{y}.
\item Extracts all rows and columns within that peak window.
\item Returns the mean of all numeric columns (and first value of
non-numeric columns) within the peak window.
}
}
\examples{
parvo_data <- read_parvo(example_epl("parvo_ramp"))$data

## Find peak 30-second VO2
find_peaks(parvo_data, x = "TIME", y = "VO2", span = 30)

## search within specific time range
find_peaks(parvo_data, x = "TIME", y = "VO2", span = 60, between = c(1500, 2000))

}
