% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_epl.R
\name{palette_epl}
\alias{palette_epl}
\title{Custom EPL colour palette}
\usage{
palette_epl(call = NULL)
}
\arguments{
\item{call}{A character or numeric vector specifying either the name or the
number in order of colours to return.}
}
\value{
Named or unnamed character vector of hex colours.
}
\description{
Custom EPL colour palette
}
\examples{
\dontrun{
scales::show_col(palette_epl())
scales::show_col(palette_epl(2))
scales::show_col(palette_epl(c("red", "orange")))
}

}
\seealso{
\code{\link[=theme_epl]{theme_epl()}} \code{\link[=scale_colour_epl]{scale_colour_epl()}}
}
