% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_epl.R
\name{palette_parvo}
\alias{palette_parvo}
\title{Custom Parvo colour palette}
\usage{
palette_parvo(...)
}
\arguments{
\item{...}{A character string specifying the named colour(s) to return.}
}
\value{
Named character vector of hex colours.
}
\description{
Custom Parvo colour palette
}
\examples{
\dontrun{
scales::show_col(palette_parvo())
scales::show_col(palette_parvo("VO2", "VCO2", "RER", "VEVO2"))
}

}
\seealso{
\code{\link[=theme_epl]{theme_epl()}} \code{\link[=scale_colour_epl]{scale_colour_epl()}} \code{\link[=palette_epl]{palette_epl()}}
}
