% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parvo_ramp}
\alias{parvo_ramp}
\title{Parvo 30-sec binned export of maximal graded cycling assessment}
\format{
A raw .CSV file with metadata and a primary data table with 15
columns and 91 rows.
}
\description{
A .CSV file exported from Parvo Medics TrueOne 2400 metabolic
cart with 30-sec binned average ventilation and gas exchange data for a
submaximal graded step test and maximal incremental ramp to V̇O~2~peak.
}
\details{
Session records a submaximal graded step test starting at 100 W and
increasing by 50 W per 5-min stage to maximum task tolerance. Followed
by a maximal incremental ramp to V̇O~2~.

Heart rate data are not valid.
}
