% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parvo.R
\name{read_parvo}
\alias{read_parvo}
\title{Read Parvo Exported CSV Data}
\usage{
read_parvo(file_path, time_column = "TIME", add_timestamp = FALSE)
}
\arguments{
\item{file_path}{The file path as a character string, including \emph{.csv} or
\emph{.xlsx} file type.}

\item{time_column}{A character string for the time column in the data table.
\emph{Defaults} to \code{time_column = "TIME"}. Used to detect the location of the data
table.}

\item{add_timestamp}{A logical to add a "timestamp" column to the data table
with date-time values, useful for synchronisation with other recordings by
time of day. Precise to ± 0.5 seconds.}
}
\value{
A list with three \link[tibble:tibble-package]{tibbles}.
\itemize{
\item \code{parvo$data} contains the data table.
\item \code{parvo$details} contains the file metadata.
\item \code{parvo$events} contains manual event inputs.
}
}
\description{
Read \emph{.CSV} or \emph{.xlsx} (but not \emph{.XLS}) files exported by Parvo Medics
TrueOne 2400 and return recorded data table and file metadata.
}
\details{
This function can only parse \emph{.CSV} files exported directly from a Parvo
metabolic cart. Obsolete \emph{.XLS} exported file format cannot be read and
must be re-saved as \emph{.xlsx} before reading with this function.
}
