% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tymewear.R
\name{read_tymewear}
\alias{read_tymewear}
\alias{read_tymewear.live}
\alias{read_tymewear.post}
\title{Read Tymewear Exported CSV Data}
\usage{
read_tymewear(file_path, ...)

\method{read_tymewear}{live}(file_path, ...)

\method{read_tymewear}{post}(file_path, ...)
}
\arguments{
\item{file_path}{The file path as a character string, including \emph{.csv}
file type.}

\item{...}{Additional arguments passed to \code{read_tymewear}.}
}
\value{
A list with two \link[tibble:tibble-package]{tibbles}.
\itemize{
\item \code{tymelive$data} contains the data table.
\item \code{tymelive$details} contains the file metadata.
}
}
\description{
Read \emph{"Live Processed CSV"} or \emph{"Post Processed CSV"} files exported from
\emph{Tymewear VitalPro} and return a list of two data frames with recorded data
and file details
}
\details{
This generic function detects \emph{Tymewear} exported file type by searching the
top of the file for recognisable column data belonging to either
\emph{"tymewear.live"} or \emph{"tymewear.post"}, and will read and import the
resulting file.
}
\examples{
## retrieve example tymewear file
file_path <- example_epl("tymewear_live")

tyme <- read_tymewear(file_path)
tyme

}
