% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_epl.R
\name{scale_colour_epl}
\alias{scale_colour_epl}
\alias{scale_color_epl}
\alias{scale_fill_epl}
\title{Scales for custom EPL palette}
\usage{
scale_colour_epl(..., aesthetics = "colour")

scale_color_epl(..., aesthetics = "color")

scale_fill_epl(..., aesthetics = "fill")
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::discrete_scale()}.}

\item{aesthetics}{A character vector with aesthetic(s) passed to
\code{ggplot2::discrete_scale()}. Default is \code{"colour"}.}
}
\value{
A ggplot2 scale object.
}
\description{
Scales for custom EPL palette
}
\seealso{
\code{\link[=palette_epl]{palette_epl()}} \code{\link[=theme_epl]{theme_epl()}}
}
