% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_epl.R
\name{theme_epl}
\alias{theme_epl}
\title{Custom \code{{epl}} ggplot2 theme}
\usage{
theme_epl(
  base_size = 14,
  base_family = "sans",
  border = c("partial", "full"),
  ink = "black",
  paper = "white",
  accent = "#0080ff",
  ...
)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family.}

\item{border}{Define either a \emph{partial} or \emph{full} border around plots.}

\item{ink}{Colour for text and lines. Default is \emph{"black"}.}

\item{paper}{Background colour. Default is \emph{"white"}.}

\item{accent}{Accent colour for highlights. Default is \emph{"#0080ff"}.}

\item{...}{Additional arguments to add to \link[ggplot2:theme]{theme()}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2} object.
}
\description{
A \link[ggplot2:ggplot2-package]{ggplot2} theme for display.
}
\details{
\itemize{
\item \code{axis.title = element_text(face = "bold")} by default. Modify to \emph{"plain"}.
\item \code{panel.grid.major} & \code{panel.grid.major} set to blank. Modify to
\verb{= element_line()} for visible grid lines.
\item \code{legend.position = "top"} by default. Modify \code{"none"} to remove legend
entirely.
\item \code{border = "partial"} uses \code{panel.border = element_blank()} and
\code{axis.line = element_line()}.
\item \code{border = "full"} uses \verb{panel.border = element_rect(colour = "black",}
\verb{linewidth = 1)} and \code{axis.line = element_line()}.
\item \code{base_family = "sans"} by default. \code{"Merriweather Sans"} is a nice
alternative font which can be installed from
\url{https://fonts.google.com/specimen/Merriweather+Sans}.
}
}
\examples{
\dontrun{
library(ggplot2)

## set theme for the current script
theme_set(theme_epl())

## plot example data
read_tymewear(example_epl("tymewear_live"))$data |>
    ggplot() +
    aes(x = time) +
    scale_colour_epl(name = NULL) +
    ylab("Ventilation Measures") +
    scale_x_continuous(
        name = "Time (mm:ss)",
        breaks = breaks_timespan(),
        labels = format_hmmss
    ) +
    geom_line(aes(y = br, colour = "BR")) +
    geom_line(aes(y = ve, colour = "VE"))
}

}
\seealso{
\code{\link[=palette_epl]{palette_epl()}} \code{\link[=scale_colour_epl]{scale_colour_epl()}}
}
