% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tymewear_live}
\alias{tymewear_live}
\title{Tymewear breath-by-breath sample export of cycling interval workout}
\format{
A raw .csv file with metadata and a primary data table with 9
columns and 1083 rows.
}
\description{
A .csv file exported from Tymewear VitalPro chest strap with
breath-by-breath export of ventilation data for a 5x 5-minute cycling
interval workout.
}
\details{
Session records a 5x 5-min cycling interval workout with the participant
alternating body position between upright training position (\emph{"UP"}) and
aerodynamic racing position (\emph{"RP"}) each work bout. Work bout Borg RPE
were between 14-15.

This session was recorded simultaneously with both Parvo and Tymewear. The
Parvo file start time is accurate to within 1 sec (± 0.5 sec precision).
The Tymewear "live" sample timestamps accurate, however other timestamps
in the "live" and "post" file metadata are \emph{NOT} accurate due to a
current issue in the recording app.

Heart rate, cycling power, cadence, and speed are exported from the Monarch
cycle ergometer in the Tymewear "live" file in wide format at the bottom
of the file.
}
