% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tymewear_post}
\alias{tymewear_post}
\title{Tymewear time series and breath-by-breath sample export of cycling
interval workout}
\format{
A raw .csv file with metadata and a primary data table with 31
columns and 2942 rows.
}
\description{
A .csv file exported from Tymewear VitalPro chest strap with
1 Hz time series and breath-by-breath export of ventilation data for a
5x 5-minute cycling interval workout.
}
\details{
Session records a 5x 5-min cycling interval workout with the participant
alternating body position between upright training position (\emph{"UP"}) and
aerodynamic racing position (\emph{"RP"}) each work bout. Work bout Borg RPE
were between 14-15.

This session was recorded simultaneously with both Parvo and Tymewear. The
Parvo file start time is accurate to within 1 sec (± 0.5 sec precision).
The Tymewear "post" start time is \emph{NOT} accurate due to a current issue
in the recording app. Sample timestamps in the Tymewear "live" file are
accurate.

The Tymewear "post" file contains two data tables: on the right is time
series data for the ventilatory and performance variables exported at
1 Hz (1 second samples). On the left is breath-by-breath samples for the
ventilation data.

Heart rate, cycling power, cadence, and speed are exported from the Monarch
cycle ergometer at 1 Hz in the time series columns.
}
