% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{The 'visualize' step of the ExplodeLayout(EL) algorithm.}
\usage{
visualize(coordinatesFile, edgelistFile, proj_name = "defaultName",
  radius = 0, nodeSize = 3, entityDegree = FALSE, labelSize = 0,
  edgeThickness = 0.1, edgeWeight = FALSE, clusterColor = TRUE,
  picFormat = "png")
}
\arguments{
\item{coordinatesFile}{a dataframe containing coordinates of all vertices for networks with radius
ranging from 0.1 to 2.0}

\item{edgelistFile}{a datafram containing three columns representing all edges in the network and
their weight}

\item{radius}{a numeric that represents the radius of the network user would like to visualize. It should be
a value in 0.1, 0.2 ... 2.0. If the user wants to output all the possible graphs, set radius equal to 0.}

\item{nodeSize}{a numeric represents the size of vertices in the network visualization}

\item{entityDegree}{a boolean variable specifying whether the vertice size should reflect its degree(vertice with more edges would be larger
and vice versa)}

\item{labelSize}{a numeric represents the size of vertices label. If user prefers not showing the label, set it as 0.}

\item{edgeThickness}{a numeric specifying the thickness of edges in the output graphs}

\item{edgeWeight}{a boolean variable specifying whether to show the weight of edges. If disabled, all edges in the output graph have the same thickness.}

\item{clusterColor}{a boolean variable specifying whether to show clusters using different colors or just give the same color to the whole network}

\item{picFormat}{a string specifying the graph file format, currently we support 'png', 'svg', and 'jpeg'}

\item{projName}{a string which is the project name or data name specified by user. This name will be used in saved export files.}
}
\description{
This function takes as input the coordinates file and edgelist file from previous step and visuallize
the networks generated from EL algorithm in saved local graph files.
}
\details{
Input:
The input of the visualize function includes mainly two data sets, which should be the output of the
previous 'search' step: 1. A 'coordinates' data containing the coordinates of all vertices in networks
with different radius(a parameter in EL algorithm). The radius range is from 0.1 to 2.0 with step of 0.1.
2. An 'edgelist' file which containing all the edges in the network and their weight.

Please note: This 'visualize' function is designed to visualize networks layouts generated using the
ExplodeLayout algorithm. The input data sets should always be the output of the previous 'search' step.

Output:
This function will generate network of particular radius(between 0.1 and 2.0) or all of them
at once depending on user's need. Plus, the initial layout of the network would always be saved
as a comparison to the 'exploded' networks. User may want to refer to the CCS scores saved in
the 'statList' file generated previously to determine the radius they'd like to set.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5543384/}
}
