% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epl.R
\docType{package}
\name{epl}
\alias{epl}
\alias{epl-package}
\title{\pkg{epl} The R package to implement the ExplodeLayout algorithm for network visuallizatoin.}
\description{
The epl package provide the powerful visuallization tool to help analyze networks.
}
\details{
The current version would enable users to apply the ExplodeLayout algorithm to any bipartite
or unipartite network.

Users have to prepare their data into particular format before calling the layout function.

In the coming version, the package will integrate more tools such as help user build the required
format, etc.
}
