% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explodeLayout.R
\name{explodelayout}
\alias{explodelayout}
\title{Test of explodelayout.}
\usage{
explodelayout(data_name = "training_sample", labelOn = "no")
}
\arguments{
\item{data_name}{the project name.}

\item{labelOn}{labels of nodes on or not. should be 'yes' or 'no'}
}
\description{
'explodelayout' reads in the network and nodelist file,
layout graphs using the ExplodeLayout algorithm with
different radius and export the results in the local files.
}
\examples{
explodelayout('training_sample', 'no')

User should first create two folders in the current
working directory. Two folders are named as 'Data'
and 'Results'. Users should put their network file and
nodelist file in the 'Data' folder. Output of the
function will be seen in the results folder.


}
