% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tymewear.R
\name{read_tymelive}
\alias{read_tymelive}
\alias{read_tymewear}
\title{Read Tymewear Exported CSV Data}
\usage{
read_tymelive(file_path)
}
\arguments{
\item{file_path}{The file path as a character string, including \emph{.csv}
file type.}
}
\value{
A list with two \link[tibble:tibble-package]{tibbles}.
\itemize{
\item \code{tymelive$data} contains the data table.
\item \code{tymelive$details} contains the file metadata.
}
}
\description{
Read from exported \emph{"Live Processed CSV"} or \emph{"Post Processed CSV"} files
and return recorded data table and file metadata.
}
