% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_outliers.R
\name{replace_outliers}
\alias{replace_outliers}
\title{Replace Local Outliers}
\usage{
replace_outliers(x, method = c("median", "NA"), width, t0 = 3)
}
\arguments{
\item{x}{A numeric vector.}

\item{method}{A character string indicating how to handle replacement
(see \emph{Details} for more on each method):
\describe{
\item{\code{"median"}}{Replaces outliers with the median within a locally
centred window defined by either \code{width} (the \emph{default}).}
\item{\code{"NA"}}{Replaces outliers with \code{NA}.}
}}

\item{width}{An integer defining the sample window in which to detect local
outliers. Where \verb{window = -width < idx < width}.}

\item{t0}{An integer for the local outlier threshold. \emph{Default} \code{t0 = 3}
(Pearson's rule).}
}
\value{
A numeric vector of filtered data.
}
\description{
Detects local outliers in vector data with a Hampel filter using median
absolute deviation (MAD), and replaces with \code{NA} or the local median value.
}
\details{
The default \code{method = "median"} will replace outliers with the local median
value, as in \code{\link[pracma:hampel]{pracma::hampel()}}. Otherwise, \code{method = "NA"} will
replace outliers with \code{NA}.

This function can be run on numeric vectors with missing \code{NA} values. \code{NA}
values in \code{x} are excluded from processing and restored in the returned
vector.

A high \code{t0} threshold makes the outlier filter more forgiving, a low one
will declare more points to be outliers. \code{t0 = 3} (the \emph{default})
corresponds to Pearson's 3 sigma edit rule, \code{t0 = 0} to Tukey's median
filter.
}
\examples{
set.seed(8421)
x <- numeric(1024)
z <- rnorm(1024)
x[1] <- z[1]
for (i in 2:1024) {
    x[i] <- 0.4*x[i-1] + 0.8*x[i-1]*z[i-1] + z[i]
}
x[150:200] <- NA ## generate NA values
y <- replace_outliers(x, width = 20, method = "median")
ind <- which(x != y) ## identify outlier indices
outliers <- x[ind] ## identify outlier values

\dontrun{
plot(1:1024, x, type = "l")
points(ind, outliers, pch = 21, col = "darkred")
lines(y, col = "blue")
}

}
\seealso{
\code{\link[pracma:hampel]{pracma::hampel()}}
}
