% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{The 'visualize' step of the ExplodeLayout(EL) algorithm.}
\usage{
visualize(coordinatesFile, edgelistFile, proj_name = "defaultName",
  radius = 0, nodeSize = 3, entityDegree = FALSE, labelSize = 0,
  edgeThickness = 0.1, edgeWeight = FALSE, clusterColor = TRUE,
  picFormat = "png")
}
\arguments{
\item{coordinatesFile}{A dataframe containing coordinates of all nodes for networks with radius
ranging from 0.1 to 2.0}

\item{edgelistFile}{A dataframe containing three columns: node-1, node-2, and edgeweight.}

\item{radius}{A numeric specifying the radius (ranging from 0.1-2.0) of the ExplodeLayout circle. For all
radius ranging from 0.1-2.0, set radius=0. Default = 3.}

\item{nodeSize}{An integer specifying the size of nodes. Default = 3.}

\item{entityDegree}{A boolean variable specifying whether the nodes should be sized by its degree (number
of edges connected to that node). Default = FALSE.}

\item{labelSize}{An integer specifying the size of the node labels. For no labels, set labelSize=0. Default = 0.}

\item{edgeThickness}{A numeric specifying the thickness of edges in the visualization. For weighted networks,
this value will be multiplied by the edge weights provided, to generate the final edge weight. Default = 0.1.}

\item{edgeWeight}{A boolean variable specifying whether to display the edge weight (using line thickness) in
weighted networks. If edgeWeight = FALSE, then all edge thicknesses will be equal. Default = FALSE.}

\item{clusterColor}{A boolean variable specifying whether to show clusters using different colors or a uniform
color. Default clusterColor = TRUE.}

\item{picFormat}{A string specifying the network layout output with values 'png', 'svg', and 'jpeg', Default = 'png'}

\item{projName}{A string specifying the project name used to export files. Default projName = 'defaultName'}
}
\description{
'This function takes as input the (1) coordinate file, and (2) edgelist file for each of the exploded
networks generated from the 'search' function. The visualize function saves the exploded networks in
local folders named by the value of the explode radius (e.g., 2).
}
\details{
Input:

(1) A 'coordinates' data object containing the coordinates of all nodes in network with different
radiuses. The default radius range is 0.1 to 2.0, with increments of 0.1.

(2) An 'edgelist' data object containing all edges connecting pairs of nodes in the network,
and their weight (if the original network is unweighted, the weight of all edges should be 1).

Please note: This 'visualize' function is designed to visualize networks layouts generated using the
ExplodeLayout algorithm. Please use the output of the 'search' step as input for the 'visualize' function.

Output:

The 'visualize' function will generate a network of a particular radius (default between 0.1 - 2.0, at 0.1 increments),
or all of them based on parameters specified by the user. Users may refer to the CCS scores saved in the
'statList' file to select a network based on the CCS scores generated from the search algorithm in the
'search' function.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5543384/}
}
